nlp_service/requirements_analysis.txt
psycopg2-binary==2.9.9
pandas==2.2.0
matplotlib==3.8.2
seaborn==0.13.1
numpy==1.26.3
python-dotenv==1.0.0
requests==2.31.0
